fix PIL/imageio CI error

+++ tests/pytorch_backend/test_pytorch_learner_backend.py
@@ -24,12 +24,14 @@ class TestUtils(unittest.TestCase):
 
     def test_write_chip(self):
         with get_tmp_dir() as tmp_dir:
-            chip = np.random.randint(0, 256, size=(100, 100, 3))
+            chip = np.random.randint(
+                0, 256, size=(100, 100, 3), dtype=np.uint8)
             path = join(tmp_dir, 'test.png')
             write_chip(chip, path)
             np.testing.assert_array_equal(np.array(Image.open(path)), chip)
 
-            chip = np.random.randint(0, 256, size=(100, 100, 8))
+            chip = np.random.randint(
+                0, 256, size=(100, 100, 8), dtype=np.uint8)
             path = join(tmp_dir, 'test.npy')
             write_chip(chip, path)
             np.testing.assert_array_equal(np.load(path), chip)
